Remove unused pages and clean code for wed development
