+++ index.android.js
@@ -1,30 +1,46 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
-
 import React, { Component } from 'react';
 import {
-  AppRegistry,
   StyleSheet,
   Text,
-  View
+  View,
+  ListView,
+  Navigator,
+  TouchableHighlight,
+  Image
 } from 'react-native';
+import NavigationBar from 'react-native-navbar'
 
-class Project extends Component {
+class Search extends Component {
+  navigate(routeName) {
+    this.props.navigator.push({
+      name: routeName
+      // passProps: {name: routeName},
+    })
+  }
+
   render() {
     return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to React Native!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit index.android.js
-        </Text>
-        <Text style={styles.instructions}>
-          Shake or press menu button for dev menu
-        </Text>
+      <View>
+        <NavigationBar
+            title={{ title:  'TipTap!' , tintColor:  'black' , }}
+            leftButton={{ title: 'Off', tintColor: 'black', handler: this.navigate.bind(this, "main")} }
+            style={{ backgroundColor:  "#D3D3D3" , }}
+            statusBar={{ tintColor:  "white" , }}
+        />
+        <View style={styles.container}>
+          <Text style={styles.welcome}>
+            Searching!
+          </Text>
+
+          <Image
+            style={{
+              width:  300 ,
+              height:  200 ,
+            }}
+            resizeMode={ "contain" }
+            source={{uri:'http://i.imgur.com/jVXo2FL.png'}}
+            />
+        </View>
       </View>
     );
   }
@@ -33,20 +49,77 @@ class Project extends Component {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+    alignItems: 'stretch',
+    backgroundColor: 'white',
+    marginTop: 24,
     justifyContent: 'center',
     alignItems: 'center',
-    backgroundColor: '#F5FCFF',
   },
   welcome: {
     fontSize: 20,
     textAlign: 'center',
     margin: 10,
   },
-  instructions: {
+  title: {
+    marginTop: 250,
+    fontSize: 20,
+    marginBottom: 8,
     textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
+  },
+  bButton: {
+    backgroundColor: '#007399',
+    color: 'white',
+    textAlign: 'left',
+    marginTop: 0,
+    fontSize: 40,
+    width: 55,
+    fontWeight: 'bold',
+  },
+  nav: {
+    justifyContent: 'flex-start',
+    width: 378,
+    height: 50,
+    backgroundColor: '#007399',
+    flexDirection: 'row',
+  },
+  navtitle: {
+    fontFamily: 'Helvetica',
+    marginTop: 15,
+    marginLeft: 74,
+    fontSize: 20,
+    color: 'white',
+    letterSpacing: 14,
+  },
+    title: {
+    marginTop: 250,
+    fontSize: 20,
+    marginBottom: 8,
+    textAlign: 'center',
+  },
+  bButton: {
+    backgroundColor: '#007399',
+    color: 'white',
+    textAlign: 'left',
+    marginTop: 0,
+    fontSize: 40,
+    width: 55,
+    fontWeight: 'bold',
+  },
+  nav: {
+    justifyContent: 'flex-start',
+    width: 378,
+    height: 50,
+    backgroundColor: '#007399',
+    flexDirection: 'row',
+  },
+  navtitle: {
+    fontFamily: 'Helvetica',
+    marginTop: 15,
+    marginLeft: 74,
+    fontSize: 20,
+    color: 'white',
+    letterSpacing: 14,
   },
 });
 
-AppRegistry.registerComponent('Project', () => Project);
+export default Search;
