+++ index.ios.js
@@ -16,10 +16,8 @@ import {
 } from 'react-native'
 
 import Main from './main'
-import Error from './error'
 import Registration from './registration'
 import Active from './active'
-import Search from './search'
 
 class Project extends Component {
   render() {
@@ -35,14 +33,8 @@ class Project extends Component {
     if(route.name === 'main') {
       return <Main navigator={navigator} {...route.passProps} />
     }
-    if(route.name === 'error') {
-      return <Error navigator={navigator} {...route.passProps} />
-    }
     if(route.name === 'active') {
       return <Active navigator={navigator} {...route.passProps} />
-    }
-    if(route.name === 'search') {
-      return <Search navigator={navigator} {...route.passProps} />
     }
     if(route.name === 'registration') {
       return <Registration navigator={navigator} {...route.passProps} />
